Move request extractors into salvo_core
