+++ crates/core/src/extract/parameter.rs
@@ -1,0 +1,12 @@
+//! Request parameter extractors.
+
+#[cfg(feature = "cookie")]
+mod cookie;
+#[cfg(feature = "cookie")]
+pub use cookie::CookieParam;
+mod header;
+pub use header::HeaderParam;
+mod path;
+pub use path::PathParam;
+mod query;
+pub use query::QueryParam;
