+++ crates/core/src/extract/parameter/cookie.rs
@@ -1,16 +1,14 @@
 use std::fmt::{self, Debug, Display, Formatter};
 use std::ops::{Deref, DerefMut};
 
-use salvo_core::Depot;
-use salvo_core::extract::{Extractible, Metadata};
-use salvo_core::http::{ParseError, Request};
-use salvo_core::serde::from_str_val;
 use serde::{Deserialize, Deserializer};
 
-use crate::endpoint::EndpointArgRegister;
-use crate::{Components, Operation, Parameter, ParameterIn, ToSchema};
+use crate::Depot;
+use crate::extract::{Extractible, Metadata};
+use crate::http::{ParseError, Request};
+use crate::serde::from_str_val;
 
-/// Represents the parameters passed by Cookie.
+/// Extracts a parameter from a request cookie.
 pub struct CookieParam<T, const REQUIRED: bool = true>(Option<T>);
 impl<T> CookieParam<T, true> {
     /// Consumes self and returns the value of the parameter.
@@ -140,26 +138,10 @@ where
     }
 }
 
-impl<T, const R: bool> EndpointArgRegister for CookieParam<T, R>
-where
-    T: ToSchema,
-{
-    fn register(components: &mut Components, operation: &mut Operation, arg: &str) {
-        let parameter = Parameter::new(arg)
-            .location(ParameterIn::Cookie)
-            .description(format!("Get parameter `{arg}` from request cookie."))
-            .schema(T::to_schema(components))
-            .required(R);
-        operation.parameters.insert(parameter);
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use assert_json_diff::assert_json_eq;
+    use crate::test::TestClient;
     use http::header::HeaderValue;
-    use salvo_core::test::TestClient;
-    use serde_json::json;
 
     use super::*;
 
@@ -293,30 +275,5 @@ mod tests {
         let result =
             CookieParam::<String, false>::extract_with_arg(&mut req, &mut depot, "param").await;
         assert_eq!(result.unwrap().0.unwrap(), "param");
-    }
-
-    #[test]
-    fn test_cookie_param_register() {
-        let mut components = Components::new();
-        let mut operation = Operation::new();
-        CookieParam::<String, false>::register(&mut components, &mut operation, "arg");
-
-        assert_json_eq!(
-            operation,
-            json!({
-                "parameters": [
-                    {
-                        "name": "arg",
-                        "in": "cookie",
-                        "description": "Get parameter `arg` from request cookie.",
-                        "required": false,
-                        "schema": {
-                            "type": "string"
-                        }
-                    }
-                ],
-                "responses": {}
-            })
-        )
     }
 }
