+++ crates/core/src/extract/parameter/header.rs
@@ -1,15 +1,13 @@
 use std::fmt::{self, Debug, Display, Formatter};
 use std::ops::{Deref, DerefMut};
 
-use salvo_core::Depot;
-use salvo_core::extract::{Extractible, Metadata};
-use salvo_core::http::{ParseError, Request};
 use serde::{Deserialize, Deserializer};
 
-use crate::endpoint::EndpointArgRegister;
-use crate::{Components, Operation, Parameter, ParameterIn, ToSchema};
+use crate::Depot;
+use crate::extract::{Extractible, Metadata};
+use crate::http::{ParseError, Request};
 
-/// Represents the parameters passed by header.
+/// Extracts a parameter from a request header.
 pub struct HeaderParam<T, const REQUIRED: bool = true>(Option<T>);
 impl<T> HeaderParam<T, true> {
     /// Consumes self and returns the value of the parameter.
@@ -144,26 +142,10 @@ where
     }
 }
 
-impl<T, const R: bool> EndpointArgRegister for HeaderParam<T, R>
-where
-    T: ToSchema,
-{
-    fn register(components: &mut Components, operation: &mut Operation, arg: &str) {
-        let parameter = Parameter::new(arg)
-            .location(ParameterIn::Header)
-            .description(format!("Get parameter `{arg}` from request headers."))
-            .schema(T::to_schema(components))
-            .required(R);
-        operation.parameters.insert(parameter);
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use assert_json_diff::assert_json_eq;
+    use crate::test::TestClient;
     use http::header::HeaderValue;
-    use salvo_core::test::TestClient;
-    use serde_json::json;
 
     use super::*;
 
@@ -297,30 +279,5 @@ mod tests {
         let result =
             HeaderParam::<String, false>::extract_with_arg(&mut req, &mut depot, "param").await;
         assert_eq!(result.unwrap().0.unwrap(), "param");
-    }
-
-    #[test]
-    fn test_header_param_register() {
-        let mut components = Components::new();
-        let mut operation = Operation::new();
-        HeaderParam::<String, false>::register(&mut components, &mut operation, "arg");
-
-        assert_json_eq!(
-            operation,
-            json!({
-                "parameters": [
-                    {
-                        "name": "arg",
-                        "in": "header",
-                        "description": "Get parameter `arg` from request headers.",
-                        "required": false,
-                        "schema": {
-                            "type": "string"
-                        }
-                    }
-                ],
-                "responses": {}
-            })
-        )
     }
 }
