+++ crates/core/src/extract/payload.rs
@@ -1,0 +1,8 @@
+//! Request body extractors.
+mod file;
+mod form;
+mod json;
+
+pub use file::{FormFile, FormFiles};
+pub use form::FormBody;
+pub use json::JsonBody;
